refact: move white label domains out of environment variables
