+++ pages/_document.js
@@ -142,7 +142,6 @@ export default class IntlDocument extends Document {
       'CAPTCHA_PROVIDER',
       'DISABLE_MOCK_UPLOADS',
       'LEDGER_SEPARATE_TAXES_AND_PAYMENT_PROCESSOR_FEES',
-      'WHITELABEL_DOMAINS',
     ]);
   }
