+++ server/content-security-policy.js
@@ -1,7 +1,7 @@
 const mergeWith = require('lodash/mergeWith');
 const { kebabCase, omit } = require('lodash');
 
-const { WHITELABEL_DOMAINS } = require('../lib/whitelabel.ts');
+import { WHITELABEL_DOMAINS } from '../lib/constants/whitelabel-providers';
 const env = process.env.OC_ENV;
 
 const SELF = "'self'";
